Move escape venv to conftest, solve some sca issues
