+++ src/conan_app_launcher/ui/views/about/about_text.html
@@ -1,6 +1,7 @@
     <!DOCTYPE html>
     <html lang="en">
-        <head title="About">
+        <head>
+            <title>About</title>
             <style type="text/css">a:link {color: #6699CC;}</style>
         </head>    
     <body style="font-weight:400; font-style:normal;">
